Modularized range filter package
