+++ src/main/java/dk/sdu/se_f22/productmodule/management/IProductManager.java
@@ -17,5 +17,4 @@ public interface IProductManager {
     void reparse();
 
     ArrayList<Product> getAllProducts();
-
 }
