--- src/main/java/dk/sdu/se_f22/sortingmodule/range/internalfilter/KeepOrNotEnum.java
@@ -1,5 +1,0 @@
-package dk.sdu.se_f22.sortingmodule.range.internalfilter;
-
-public enum KeepOrNotEnum {
-    KEEP, REMOVE, UNKNOWN
-}
