+++ src/main/java/dk/sdu/se_f22/sortingmodule/range/rangefilter/RangeFilterCreator.java
@@ -1,5 +1,6 @@
 package dk.sdu.se_f22.sortingmodule.range.rangefilter;
 
+import dk.sdu.se_f22.sharedlibrary.SearchHits;
 import dk.sdu.se_f22.sortingmodule.range.dbrangefilter.ReadRangeFilterInterface;
 import dk.sdu.se_f22.sortingmodule.range.dbrangefilter.CreateRangeFilterInterface;
 import dk.sdu.se_f22.sortingmodule.range.dbrangefilter.DBRangeFilter;
@@ -8,7 +9,10 @@ import dk.sdu.se_f22.sortingmodule.range.dbrangefilter.DBRangeFilterReader;
 import dk.sdu.se_f22.sortingmodule.range.exceptions.InvalidFilterException;
 import dk.sdu.se_f22.sortingmodule.range.exceptions.InvalidFilterIdException;
 
-class RangeFilterCreator {
+import java.util.Collection;
+import java.util.List;
+
+public class RangeFilterCreator  implements RangeFilterInterface {
     private CreateRangeFilterInterface dbRangeFilterCreator;
     private ReadRangeFilterInterface dbReader;
 
@@ -32,7 +36,7 @@ class RangeFilterCreator {
         return true;
     }
 
-    public InternalFilter createInternalFilter(UserInputtedRangeFilter filterCheck) throws InvalidFilterIdException {
+    public InternalFilter createInternalFilter(RangeFilter filterCheck) throws InvalidFilterIdException {
         //check if the filter exists in the database
         //if it doesn't, or min, max is invalid
 //        return null;
@@ -43,5 +47,29 @@ class RangeFilterCreator {
         } else {
             return null;
         }
+    }
+
+    /**This is the method that filters the products in the searchHits based on the filters given.
+     *
+     * @param rangeFilters The rangefilters to use for filtering the search hits, they must be in accordance with the filters stored in our DB.
+     *                     See {@link ReadRangeFilterInterface} for details on getting active/valid filters.
+     * @return The {@link SearchHits} object that was given as input, but where the products Colloction have been filtered
+     * using the filters specified in rangeFilters param.
+     */
+    public SearchHits filterResults(SearchHits searchHits, List<RangeFilter> rangeFilters) throws InvalidFilterIdException {
+        Collection productHits = searchHits.getProducts();
+
+        for (RangeFilter rangeFilter : rangeFilters) {
+            InternalFilter iFilter = this.createInternalFilter(rangeFilter);
+
+            if (iFilter == null) {
+                continue;
+            }
+
+
+            productHits = iFilter.useFilter(productHits);
+        }
+
+        return searchHits;
     }
 }
