+++ src/main/java/dk/sdu/se_f22/sortingmodule/range/rangefilter/RangeFilterInterface.java
@@ -15,5 +15,5 @@ public interface RangeFilterInterface {
      *  @param rangeFilters The rangefilters to use for filtering the search hits
      * @return
      */
-    SearchHits filterResults(SearchHits searchHits, List<UserInputtedRangeFilter> rangeFilters) throws InvalidFilterIdException;
+    SearchHits filterResults(SearchHits searchHits, List<RangeFilter> rangeFilters) throws InvalidFilterIdException;
 }
