+++ src/main/java/dk/sdu/se_f22/sortingmodule/range/rangefilter/internalfilter/KeepOrNotEnum.java
@@ -1,0 +1,5 @@
+package dk.sdu.se_f22.sortingmodule.range.rangefilter.internalfilter;
+
+public enum KeepOrNotEnum {
+    KEEP, REMOVE, UNKNOWN
+}
