+++ src/test/java/dk/sdu/se_f22/sortingmodule/range/Helpers.java
@@ -1,6 +1,6 @@
 package dk.sdu.se_f22.sortingmodule.range;
 
-import dk.sdu.se_f22.sortingmodule.range.rangefilter.UserInputtedRangeFilter;
+import dk.sdu.se_f22.sortingmodule.range.rangefilter.RangeFilter;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -57,8 +57,8 @@ public class Helpers {
         return attributeMap;
     }
 
-    public static List<UserInputtedRangeFilter> readRangeFiltersFromFile(String fileName){
-        List<UserInputtedRangeFilter> mockResults = new ArrayList<>();
+    public static List<RangeFilter> readRangeFiltersFromFile(String fileName){
+        List<RangeFilter> mockResults = new ArrayList<>();
 
         List<String> filters = Helpers.readFromCSV(fileName);
 
@@ -71,7 +71,7 @@ public class Helpers {
                 continue;
             }
 
-            mockResults.add(new UserInputtedRangeFilter(Integer.parseInt(filterSplit[0]), Double.parseDouble(filterSplit[1]), Double.parseDouble(filterSplit[2])));
+            mockResults.add(new RangeFilter(Integer.parseInt(filterSplit[0]), Double.parseDouble(filterSplit[1]), Double.parseDouble(filterSplit[2])));
         }
 
         return mockResults;
