--- src/test/java/dk/sdu/se_f22/sortingmodule/range/RangeFilterTest.java
@@ -1,5 +1,0 @@
-package dk.sdu.se_f22.sortingmodule.range;
-
-class RangeFilterTest {
-
-}
