+++ src/test/java/dk/sdu/se_f22/sortingmodule/range/dbrangefilter/DBRangeFilterReaderTest.java
@@ -1,5 +1,6 @@
-package dk.sdu.se_f22.sortingmodule.range;
+package dk.sdu.se_f22.sortingmodule.range.dbrangefilter;
 
+import dk.sdu.se_f22.sortingmodule.range.PopulateDBFromCsv;
 import dk.sdu.se_f22.sortingmodule.range.database.DatabaseInterface;
 import dk.sdu.se_f22.sortingmodule.range.database.MockDatabase;
 import dk.sdu.se_f22.sortingmodule.range.dbrangefilter.DBRangeFilter;
