+++ src/test/resources/dk/sdu/se_f22/sortingmodule/range/validators/emptyStrings.csv
@@ -1,0 +1,5 @@
+%
+-
+.
+/
+*
